fix MapEvents struct

+++ include/global.fieldmap.h
@@ -77,7 +77,7 @@ struct BgEvent
 
 struct MapEvents
 {
-    u32 mapObjectCount;
+    u8 mapObjectCount;
     u8 warpCount;
     u8 coordEventCount;
     u8 bgEventCount;
